도서 검색 상세페이지 평점 리뷰 기능 구현 진행

+++ src/main/java/com/bookstore/home/domain/ReviewDto.java
@@ -13,7 +13,7 @@ public class ReviewDto {
 
     private int reviewNum;
 
-    private int reviewGrade;
+    private double reviewGrade;
 
     private String memberEmail;
